refactor: extract ThinStr to its own module
